Intervals_virtual-class {intervals}	R Documentation

Class "Intervals_virtual"

Description

A virtual class from which the "Intervals" and "Intervals_full" classes derive.

Slots

.Data:

Object of class "matrix". A two-column, numeric (see below) format is required. For a valid object, no value in the first column may exceed its partner in the second column. (Note that this does permit empty interval rows, when both endpoints are of equal value and not both closed.) Only integral (though not "integer" class) endpoints are permitted if type is "Z". See the note on this point in documentation for "Intervals".

type:

Object of class "character". A one-element character vector with either "Z" or "R" is required.

Extends

Class "matrix", from data part.

Class "array", by class "matrix", distance 2.

Class "structure", by class "matrix", distance 3.

Class "vector", by class "matrix", distance 4, with explicit coerce.

Methods

close_intervals

signature(x = "Intervals_virtual")

closed

signature(x = "Intervals_virtual")

clusters

signature(x = "Intervals_virtual")

coerce

signature(from = "Intervals_virtual", to = "character")

contract

signature(x = "Intervals_virtual")

expand

signature(x = "Intervals_virtual")

head

signature(x = "Intervals_virtual")

initialize

signature(.Object = "Intervals_virtual")

interval_complement

signature(x = "Intervals_virtual")

interval_difference

signature(x = "Intervals_virtual", y = "Intervals_virtual")

interval_intersection

signature(x = "Intervals_virtual")

interval_union

signature(x = "Intervals_virtual")

is.na

signature(x = "Intervals_virtual")

open_intervals

signature(x = "Intervals_virtual")

reduce

signature(x = "Intervals_virtual")

show

signature(object = "Intervals_virtual")

t

signature(x = "Intervals_virtual")

tail

signature(x = "Intervals_virtual")

type

signature(x = "Intervals_virtual")

type<-

signature(x = "Intervals_virtual")

which_nearest

signature(from = "numeric", to = "Intervals_virtual")

which_nearest

signature(from = "Intervals_virtual", to = "numeric")

which_nearest

signature(from = "Intervals_virtual", to = "Intervals_virtual")

See Also

See the "Intervals" and "Intervals_full" classes.

[Package intervals version 0.15.5 Index]