interval_complement {intervals}	R Documentation

Compute the complement of a set of intervals

Description

Compute the complement of a set of intervals.

Usage

## S4 method for signature 'Intervals_virtual'
interval_complement(x, check_valid = TRUE)

Arguments

x
An "Intervals" or "Intervals_full" object.

check_valid
Should validObject be called before passing to compiled code? Also see interval_overlap.

Value

An object of the same class as x, compactly representing the complement of the intervals described in x.

Note

For objects of class "Intervals", closure on -Inf or Inf endpoints is set to match that of all the intervals with finite endpoints. For objects of class "Intervals_full", non-finite endpoints are left open (although in general, this package does not make a distinction between closed and open infinite endpoints).

[Package intervals version 0.15.1 Index]