intervals-package {intervals}	R Documentation

Tools for working with points and intervals

Description

Tools for working with and comparing sets of points and intervals.

Details

Index:

Intervals-class

Classes "Intervals" and "Intervals_full".

Intervals_virtual-class

Class "Intervals_virtual".

Intervals_virtual_or_numeric-class

Class union "Intervals_virtual_or_numeric".

as.matrix

Coerce endpoints to a matrix.

c

Concatenate different sets of intervals.

close_intervals

Re-represent integer intervals with open or closed endpoints.

closed

Accessor for closed slot: closure vector/matrix.

clusters

Identify clusters in a collection of positions or intervals.

contract

Contract sets.

distance_to_nearest

Compute distance to nearest position in a set of intervals.

empty

Identify empty interval rows.

expand

Expand sets.

interval_complement

Compute the complement of a set of intervals.

interval_difference

Compute set difference.

interval_included

Assess inclusion of one set of intervals with respect to another.

interval_intersection

Compute the intersection of one or more sets of intervals.

interval_overlap

Assess which query intervals overlap which targets.

interval_union

Compute the union of intervals in one or more interval matrices.

is.na

Identify interval rows with NA endpoints.

plot

S3 plotting methods for intervals objects.

reduce

Compactly re-represent the points in a set of intervals.

sgd

Yeast gene model sample data.

size

Compute interval sizes.

split

Split an intervals object according to a factor.

type

Accessor for type slot: Z or R.

which_nearest

Identify nearest member(s) in a set of intervals.

Further information is available in the following vignettes:

intervals_overview

Overview of the intervals package.

Acknowledgments

Thanks to Julien Gagneur, Simon Anders, and Wolfgang Huber for numerous helpful suggestions about the package content and code.

Author(s)

Richard Bourgon <bourgon@ebi.ac.uk>

See Also

See the genomeIntervals package in Bioconductor, which extends the functionality of this package.

[Package intervals version 0.15.5 Index]