c {intervals}	R Documentation

Combine different interval matrix objects

Description

S3 methods for concatenating sets of intervals into a single set.

Usage

## S3 method for class 'Intervals'
c(...)
## S3 method for class 'Intervals_full'
c(...)

Arguments

...
"Intervals" or "Intervals_full" objects.

Details

All objects are expected to have the same value in the type slot. If the closed slots differ for "Intervals" objects and type == "Z", the objects will be adjusted to have closed values matching that of x; if type == "R", however, then all objects must first be coerced to class "Intervals_full", with a warning. This coercion also occurs when a mixture of object types is passed in. A NULL in any argument is ignored.

Value

A single "Intervals" or "Intervals_full" object. Input objects are concatenated in their order of appearance in the the argument list.

If any input argument is not a set of intervals, list(...) is returned instead.

Note

These methods will be converted to S4 once the necessary dispatch on ... is supported.

Examples

f1 <- Intervals( 1:2, type = "Z" )
g1 <- open_intervals( f1 + 5 )

# Combining Intervals objects over Z may require closure adjustment
c( f1, g1 )

f2 <- f1; g2 <- g1
type( f2 ) <- type( g2 ) <- "R"

# Combine Intervals objects over R which have different closure requires
# coercion

h <- c( f2, g2 )

# Coercion for mixed combinations as well
c( h, g2 + 10 )

## Not run: 
# Combining different types is not permitted
c( h, g1 + 10 )

## End(Not run)

[Package intervals version 0.15.1 Index]