distance_to_nearest {intervals}	R Documentation

Compute distance to nearest position in a set of intervals

Description

For each point or interval in the from argument, compute the distance to the nearest position in the to argument.

Usage

## S4 method for signature 
## 'Intervals_virtual_or_numeric,Intervals_virtual_or_numeric'
distance_to_nearest(from, to, check_valid = TRUE)

Arguments

from
An object of appropriate type.

to
An object of appropriate type.

check_valid
Should validObject be called before passing to compiled code? Also see interval_overlap.

Value

A vector of distances, with one entry per point or interval in from. Any intervals in from which are either empty (see empty) or have NA endpoints produce a NA result.

Note

This function is now just a wrapper for which_nearest.

See Also

See which_nearest, which also returns indices for the interval or intervals (in case of ties) at the distance reported.

Examples

# Point to interval

to <- Intervals( c(0,5,3,Inf) )
from <- -5:10
plot( from, distance_to_nearest( from, to ), type = "l" ) 
segments( to[,1], 1, pmin(to[,2], par("usr")[2]), 1, col = "red" )

# Interval to interval

from <- Intervals( c(-Inf,-Inf,3.5,-1,1,4) )
distance_to_nearest( from, to )

[Package intervals version 0.15.1 Index]