interval_difference {intervals}	R Documentation

Compute set difference

Description

Compute the set difference between two objects.

Usage

## S4 method for signature 'Intervals_virtual,Intervals_virtual'
interval_difference(x, y, check_valid = TRUE)

Arguments

x
An "Intervals" or "Intervals_full" object.

y
An "Intervals" or "Intervals_full" object, with a type slot matching that of x.

check_valid
Should validObject be called on x and y before passing to compiled code? Also see interval_overlap.

Value

An object representing the subset of the integers or real line, as determined by type(x), found in x but not in y.

See Also

These methods are just wrappers for interval_intersection and interval_complement.

[Package intervals version 0.15.1 Index]