interval_intersection {intervals}	R Documentation

Compute the intersection of one or more sets of intervals

Description

Given one or more sets of intervals, produce a new set compactly representing points contained in at least one interval of each input object.

Usage

## S4 method for signature 'Intervals_virtual'
interval_intersection(x, ..., check_valid = TRUE)

## S4 method for signature 'missing'
interval_intersection(x, ..., check_valid = TRUE)

Arguments

x
An "Intervals" or "Intervals_full" object.

...
Additional objects of the same classes permitted for x.

check_valid
Should validObject be called before passing to compiled code? Also see interval_overlap.

Value

A single object representing points contained in each of the objects supplied in the x and ... arguments.

See Also

See interval_union and interval_complement, which are used to produce the results.

[Package intervals version 0.15.1 Index]