interval_union {intervals}	R Documentation

Compute the union of intervals in one or more interval matrices

Description

Compute the union of intervals in one or more interval matrices. The intervals contained in a single interval matrix object need not, in general, be disjoint; interval_union, however, always returns a matrix with sorted, disjoint intervals.

Usage

## S4 method for signature 'Intervals_virtual'
interval_union(x, ..., check_valid = TRUE)

## S4 method for signature 'missing'
interval_union(x, ..., check_valid = TRUE)

Arguments

x
An "Intervals" or "Intervals_full" object.

...
Optionally, additional objects which can be combined with x. See c.Intervals for details on mixing different types of objects.

check_valid
Should validObject be called before passing to compiled code? Also see interval_overlap.

Details

All supplied objects are combined using c and then then passed to reduce. The missing method is only to permit use of do.call with named list, since no named element will typically match x.

Value

A single object of appropriate class, compactly representing the union of all intervals in x, and optionally, in ... as well. For class "Intervals", the result will have the same closed values as x.

See Also

See reduce, which is used to produce the results.

[Package intervals version 0.15.1 Index]