split {intervals}	R Documentation

Split an intervals object according to a factor

Description

S3 and S4 methods for splitting "Intervals" or "Intervals_full" objects.

Usage

## S3 method for class 'Intervals_virtual'
split(x, f, drop = FALSE, ...)

## S4 method for signature 'Intervals_virtual'
split(x, f, drop = FALSE, ...)

Arguments

x
"Intervals" or "Intervals_full" objects.

f
Passed to split.data.frame.

drop
Passed to split.data.frame.

...
Passed to split.data.frame.

Value

A list of objects of the same class as x, split by the levels of f. Until R 2.15, special handling was not required. Subsequent changes to the base package split function required an explicit method here, but code already provided by split.data.frame was sufficient.

[Package intervals version 0.15.1 Index]